Updated a u8 in Battle Tower that causes an overflow in Multi Battles.

+++ src/battle_tower.c
@@ -1166,7 +1166,7 @@ void SetBattleFacilityTrainerGfxId(u16 trainerId, u8 tempVarId)
 {
     u32 i;
     u8 facilityClass;
-    u8 trainerObjectGfxId;
+    u16 trainerObjectGfxId;
 
     SetFacilityPtrsGetLevel();
     #ifndef FREE_BATTLE_TOWER_E_READER
